Bug#35793818 SEGFAULT in NDB_SCHEMA_OBJECT::release
Problem

Seems like there are some rare double-frees of
ndb_schema_object instances.

Suspect concurrent calls to release() and get()
where :
 - release() decrements object's use count to zero
   ...
 - get() finds object
   ...
 - release() deletes the object
 ...
 - crash on second release / later access

Solution

Design invariant seems to be that use count
for an object in the collection should always
be >= 1.  This is its value on construction.

Problem is that concurrent users can get() an
object in the process of being removed.

This part of the lifecycle should be protected
by the collection lock so that only stable states
are visible externally.

Changes
 - Add increment_use_count() precondition that
   use_count > 0
 - Collection lock extended to cover all of
   release() call
   - (decrement use count to 0, remove from collection, delete)
   - Should be safe as
     - get() already has this lock ordering (collection, object)
 - Fix type decremement -> decrement

Change-Id: I4165474b8eced1e55f4da3f46742ef7ecd25c756