Bug#37382551 Sync both schema and tables in the same interval on ndb_metadata_check enabling
When metadata change monitor detects a new database it takes until the
next check cycle for the tables in that database to be detected.

Fix by adding newly detected schema to list of schemas to also check for
tables. This should speed up detection of the tables in the new schema.

Change-Id: I81094fc231f1db4d92ededd5dadba97b4fa3a080