Bug#37372650 String buffer overflow in ndbinfo
In ndbinfo tables, all VARCHAR columns are limited to no more than
512 characters. Several ndbinfo tables include columns containing
a comma-separated list of names; this includes the primary_key column
in dictionary_tables and the "columns" column of the events and
index_columns tables.

These lists are informative. If the length of a column list
exceeds 512 characters, the list will be truncated.

Starting with this change, any individual column name in these lists
that contains a backtick, comma, or dot character will be quoted
in backticks following the usual practice of MySQL identifiers. In
the event that the list is truncated, the final name in the list
will be an unquoted string of three dots.

This patch adds a new test case ndbinfo_virtual.

Change-Id: I3101ed39a492ef62fc0602dd95cc69b801949460