Bug#37426282 innodb.ibuf_sys_tablespace_extend test failing sporadically
Description
===========
The test empirically tries to hold as many records in the change
buffer B-tree so that Innodb is required to expand the size of the
system tablespace where the change buffer is stored.
But sometimes there are not enough records in the change buffer that
force the tablespace expansion, thus test fails.

Fix
===
Improved the test such that more records are stored in the change
buffer :

- Enabled the innodb_change_buffering_debug system variable to evict
  the blocks as much possible from the the buffer pool.

- Updating more fields used in the secondary indexes

- Reduced the number of tables and connections from 4 to 2 as rather
  modifying more records in the two tables.

Also did following general test improvements :

- Converted the text printed in the log from information to warning and
  reduced the log level. This reduces the IOs in error log and it is
  quicker to search.

- Keep the default buffer pool size set by the test.

- Removed the unnecessary printing of the sys variables.

- Removed dropping the schema as we remove the entire datadir.

- Added provision of detecting the the system tablespace file size.
  If size doesn't change then test would fail.

Change-Id: Iaab92411bac8b5ecac574332086dd1bc8634dedd