Bug#35507777 MySQL Server crashes when executing query
Normally, with an unused window definition, this would be an updatable
view. But here it contains a subquery, which makes resolving during
view creation mark the view as not updateable (because we do not
eliminate subqueries during view definition and windows are not
eliminated if they contain subqueries in their ordering clauses) and
the presence of windows makes the query unmergeable and hence
unupdateable.  At UPDATE time, the window will be eliminated, but too
late.

Solution: test mergeability by checking the presence of window
functions, rather than window definitions. This lets the current query
be updateable.

Change-Id: I970add4607478d9c187097120f33e980ac598857