Bug#37492169 ndb_restore --include-stored-grants not restoring table
Problem:

On investigating an issue with restoring stored grants from a NDB
backup, it was discovered that if NDB data nodes are restarted from
`--inital` state, then the restoration of the stored
grants (--include-stored-grants) would not create nor restore the
`ndb_sql_metadata` table.

Analysis:

On looking at the `ndb_restore` code, namely the `Restore.cpp` file
where some tables are defined as system tables (`m_isSysTable =
true`), it was found that `ndb_sql_metadata` is included amongst
those. Those tables are recreated or restored from other means. A
previous commit included the restoration of the `ndb_sql_metadata` iff
option `--include-stored-grants` was given.

Solution:

This commit implements two parts, connected to each other:

- Restored the `--include-stored-grants` value switch to the code that
checks if a table is a system table and thus not to be restored. This
allows the `ndb_sql_metadata` table to NOT be deemed as system table
if the option `--include-stored-grants` is set.

- Adds skipping checks to the recreation of the table and its PRIMARY
index, so that no "already exists" error is raised, for any
table/index, and the work continues. Tests changed to reflect this
"gentle" behavior.

Change-Id: I3575ad140677bcd0b40be6c4c5cfcbb870fdd1a1