BUG#37489870 missing purge wait done logging
Log message for when PURGE BINARY LOGS was done waiting
for purger is missing. This makes it hard to tell wheter
purges are started which never completes or indefinite
wait occurs.

Fix by adding log message when wait for purger is done and also
make the log message for when starting to wait more distinct.

Change-Id: Ibd7a73f2267f6fc048bb13191565029a0e6a1799