Bug #34454572 : Clone_persist_gtid causes memory leak
Description:
============
The Clone_persist_gtid thread causes memory leak.

Analysis:
=========
The Clone_persist_gtid thread that runs in the background write GTIDs
periodically to disk table. In this process, the thd allocates memory
to store the information about the tables to be locked while flushing
GTIDs. This memory is not freed until the shutdown is called, causing
the memory consumption or slow memory leak.

Fix:
====
Release the memory that was allocated by calling the ClearForReuse()
after the GTIDs are flushed in each cycle.

Change-Id: Ib28a2c5134b5d223c48224916911199f260f52b5