Bug #37022679 - Too Much Disk Read on Startup, penalizing deployments with many tables (1M+).
SYMPTOM:
- Observing significant performance regression during server startup
in 8.0.39 compared to 5.7.44 with 1 million tables at the startup.

ROOT CAUSE:
- The design of the startup was changed in WL#8619 which includes
these two steps:
i. It reads the first page of each of the ibd file to determine
the space id of the tablespace.
ii. It validates the path in which the ibd file resides and stores
the information in DD.
The opening and reading of each ibd file's first page is taking
significant amount of time. The validation of path can be turned
off using innodb_validate_tablespace_paths=OFF.
In 5.7 tablespace id of a file was read from the redo logs which
was faster since we didn't scan through the ibd files. This design
was changed in 8.0 to support WL#8619.

FIX:
1. We open the ibd files with O_DIRECT mode for a faster access by
bypassing the OS page cache.
2. If disabling OS cache would result in an error, we would ignore
it with a newly added option to do so.
3. Instead of reading the whole page, we read just the sector size
i.e., 512 bytes. It is better than reading the 16KB pages.
The first 38 bytes from the header of the first page is suffcient
for us to know the space id of the ibd file.

Change-Id: I3ddb72b3cd0fb5089ef8976aaa1daa9b4425e031