Bug#37505746 Bugfix of Bug#37318367 in 8.0.41 contains an obvious logical issue
Description
------------
We introduced two wrapper APIs over the persistent cursor (pcursor)
through Bug#37318367 fix. They are paired up as following :
- save_previous_user_record_as_last_processed()
- restore_to_first_unprocessed()
When the restore_to_first_unprocessed() is called the cursor's relative
position is set as BTR_PCUR_AFTER in the m_rel_pos member.

In the API implementation, we handled an explicit case that is,
if cursor is restored to the same user records where it was
positioned before, then we explicitly move the cursor to the next user
record.
This explicit handling may seem legitimate but it is incorrect and not
required as well.

The reason this handling is incorrect is that the APIs for moving to
the next record expect the current record position (m_rel_pos) to be
BTR_PCUR_IS_POSITIONED, which they assert.

The reason this handling is not required is that
save_previous_user_record_as_last_processed() is called when the
cursor is pointing at SUPREMUM, and if optimistic restoration succeeds,
then the cursor still points to the SUPREMUM, the flag
BTR_PCUR_IS_POSITIONED_OPTIMISTIC is not set, and we don't have to
move the cursor forward - it's already where we wanted it to be.

Fix

Change-Id: I30e5ae719d14c48ed4e626b77435a4a4c85112e1
---
Removed the unintended code and added test cases to verify optimistic
restore of cursor without BTR_PCUR_IS_POSITIONED_OPTIMISTIC is set.

In case the flag BTR_PCUR_IS_POSITIONED_OPTIMISTIC is set, cursor
restores in one direction that is to BTR_PCUR_BEFORE therefore,
removed the debug assert checks for BTR_PCUR_AFTER.

We thank you Mr Yichang SONG at Alibaba for reporting this bug.

Change-Id: I3ba9720e9df2ee7934edcfac726987b1a3b18246