Bug#36421690 	mysql server 8.3.0 crashed at Item_ident::fix_after_pullout
Mysqld crashes while resolving a nested subquery with an ORDER BY on
a derived table column.

When an ORDER list is pushed down from an outer query block to an
inner query block, the context query block of each ORDER item should
be updated. However, this is skipped in the bug scenario because
no inner query expression is present. So the ORDER items have
invalid context query blocks after pushdown.

Later, while merging a derived table, the ORDER items are fixed to
update the dependency information. Each ORDER item's context query
block is expected to be either the outer query block or an ancestor
of the outer query block. This is not the case in the bug scenario
since the context query blocks were not updated during order list
pushdown. This results in a null-pointer dereference.

This is fixed by ensuring that all ORDER items are updated with
the correct context query block while pushing down an ORDER list,
even when there is no inner query expression.

Change-Id: I408d3a85dd13a146778442860015aeca31839a2e