Bug#37524092 Improve Api Failure handling logs + limit duration
Implement a time limit on API failure handling at the data nodes.

QMGR already logs when API failure handling is taking a long time.

This is extended with a time limit on how long a data node is
allowed to take to handle an API node failure before it is
considered a failure of the data node itself.

API failure handling involves cleanup of transaction state
in the DICT, TC, SPJ and SUMA blocks.

DICT transaction state can include schema transactions which may
take some time to complete, so a hard-coded long timeout (7 days)
is used for cases where API failure handling is blocked on DICT
processing.

For other blocks, a shorter configurable timeout is used.

When the timeout elapses, the data node with the problem will
be shutdown, which will hopefully help resolve the problem.

The short timeout is configured with a new data node configuration
parameter :

 ApiFailureHandlingTimeout
   Units : Seconds
   Where :
     0 = No limit
     1..10 = 10 seconds
     > 10 = Limit in seconds
   Default: 600

Two new tests are added to extend coverage of API failure
handling :

  testNodeRestart -n multi_apifail
    Coverage of concurrent API failure handling

  testNodeRestart -n timeout_apifail
    Coverage of data node timeout of api failure handling

Change-Id: Iefea39042cdd4ea83fb22fe2681f9e7bf7d56dba