Bug#32442772 LOG_STATUS HAS WRONG BINARY_LOG_POSITION OF GTID_EXECUTED
Description:
------------
When the server is under heavy write operation, binary_log_position
of gtid_executed shown in log_status doesn't match with the position
of that GTID events shown in the binary log file.

Analysis:
---------
A binlog commit has the following stages:
1. Flush  – Protected by the LOCK_log.
2. Sync   – At this stage, LOCK_log is released, and LOCK_sync is
            acquired.
3. Commit – Here, LOCK_sync is released, and LOCK_commit is acquired.
            During this stage, global_sid_lock is also acquired,
            and gtid_executed is updated.

When querying the log_status table, LOCK_log is acquired for the
binlog at Log_resource_binlog_wrapper::lock, and global_sid_lock is
acquired for gtid_executed at Log_resource_gtid_state_wrapper::lock.

The issue arises when a query to the log_status table is executed
after stage 1 of the binlog commit. At this point, LOCK_log has
already been released, the binlog file and position are updated, and
Log_resource_binlog_wrapper::lock proceeds. However, while the binlog
commit is still processing stage 2,
Log_resource_gtid_state_wrapper::lock is invoked and acquires
global_sid_lock. At this point, stage 3 of the binlog commit is
unable to update gtid_executed, leading to a mismatch between the
binlog position(the new values) and gtid_executed(old value) in the
log_status query.

Fix:
----
Increase the scope of the lock on the log_status table to also ensure
that the transactions that entered in the binlog commit pipeline are
completed. This adjustment will ensure that the query to the
log_status table waits until gtid_executed has been fully updated,
thereby guaranteeing consistency with the binlog position.


Change-Id: Iab0513821a0e1ffcc8a522b61447b0d9193aafe8