Bug#37436310 - Performance issue with Multi-Valued Index and ORDER BY DESC LIMIT (Reads more data than expected)
Issue:
A query having a Multi-Valued Index with an ORDER BY DESC is
slow.

Analysis:
In case of multi-valued index the key comparison to detect
end of range scan was not handled correctly which lead to
reading more number of rows in case of reverse index range
scan.

Fix:
Fixed the logic to correctly check the end of range.

Change-Id: I41e29ec5aee536a864ae78973625987f16327791