BUG#37597512: Backport BUG#37569333 into 8.0
In Group Replication dependencies between transactions are tracked
by comparing the write-sets of transactions.
If two transactions do write on the same rows they are dependent and
must be applied one after the other. If two transactions do not
write on the same rows they are independent and can be applied
concurrently.
MySQL is only able to extract the write-set from DMLs that
insert/update rows.
DDL and empty transactions do not have write-set, thence these
transactions are classified as dependent and applied sequentially.
While DDL must be applied sequentially, there is no actual reason to
force the same behaviour for empty transactions.

This patch makes possible that empty transactions are applied
concurrently with any other non dependent transaction.

Change-Id: Iffdd1f218e5b13491e138694c81f740f1200c929