Bug#37587252: Backport Bug#34348094 and Bug#36047891 into 8.0
== First bug

Bug#34348094 p_s table presented GR primary with majority after it was expelled for 12 hours

Problem
=======================================
A group running group replication with
a primary (i1) and two secondaries (i2, i3) started to have
intermittent issues because of high memory usage in the
primary group member (i1).

Secondaries (i2, i3) started to report primary (i1) as
unreachable/reachable, and the primary (i1) started to
report the secondaries (i2, i3) as unreachable/reachable.

After some instability, the secondaries (i2, i3) decided
to expel the primary (i1) and elected a new primary (i2).

Queries in the former primary (i1)
performance_schema.replication_group_members reported i1
ONLINE and PRIMARY,  i2 ONLINE and SECONDARY and i3 ONLINE
and SECONDARY for at least 12 hours until a request to
shutdown i1's MySQL

Analysis and Fix
======================================
After it happened again, we were able to trace the problem
in the Primary (i1). One of the secondaries was overloaded
and it was intermittently leaving and joining the group,
and having its connections dropped and recreated on the
primary server. In that reconnection process, the primary
got stuck when trying to create that connection,
thus blocking the XCom thread, hampering the progress of
XCom, which we know as being single threaded.

This was traced to the invocation of SSL_connect on the XCom
communication stack. It changed in WL#9852, from an
asynchronous to a synchronous form to create an uniform
way of connecting to other nodes, instead of having two
separate functions to do the same. The problem is that,
if a node is overloaded, it can fail to answer on
SSL_connect, thus leaving the connecting end blocked forever.

To tackle this issue, we created a timed_connect_ssl_msec
method, that is non-blocking and that returns in case
of a timeout, leaving the retry attempts to the caller. In
this specific case, the XCom thread when trying to reconnect
to another node.

== Second bug

Bug#36047891 SPORADICALLY GCS TIMES OUT waiting for a connection via poll to be established

Problem
======================
Many GR tests are failing on windows runs with error 'Timed out while waiting
for a connection via poll to be established'

Analysis
======================
Suppress on tests the error message that was introduced in the previous fix

Change-Id: Ib0adf4702c8e0e0f8b42cf70cd058b6dfc0faa62