BUG#37193011: crashing in sp_prepare_func_item() after               changing definition of routines.
DESCRIPTION:
============
Executing a stored routine whose definition is modified
after it is prepared, would lead to a server exit.

ANALYSIS:
=========
There is handling in place during the preparation of a
stored routine, that detects a change of definition and an
appropriate error is reported. However, if the definition is
changed post the preparation, invocation of the stored
routine would lead to a server exit.

FIX:
====
Added error handling in the Stored procedure and Stored
Function execution phase.

Note:
=====
The error handling in the SP execution phase was removed by
the patch for Bug#25398451: Segmentation fault in
prepare_inner() at sql_call.cc.
The patch for WL#9384: Prepare each DML statement once, made
sure the preparation is performed only once thus causing the
current scenario to lead to a server exit.

Change-Id: I8161b35f482565fe60951b07f0adc4fa340a02a4