Bug#37188154 Cluster/J Buffer Management errors
This patch addresses the reported issue in MySQL 8.0 and up.

A new Cluster/J test case, ByteBufferPoolTest, adds tests for using
non-default values for the Cluster/J configuration property
com.mysql.clusterj.byte.buffer.pool.sizes.

Running these tests revealed two flaws in the implementation. One was
a failure to set the buffer position back to zero in
VariableByteBufferPoolImpl.initializeGuard(); it is my understanding
that this has made all oversized buffers (i.e. larger than the largest
pooled size) unusable. Interestingly, this implies that the "cleaning"
implemented in VariableByteBufferPoolImpl is never used.

In this patch I also slightly simplified VariableByteBufferPoolImpl
by using queues.ceilingEntry(), which did not exist in the standard
library before Java 1.6.

The other problem was a very old apparent copy & paste error in DbImpl,
which could cause buffers of size 500 bytes to be improperly returned
to the queue of 1000-byte buffers. I took the extra step of redesigning
the returnBuffer() API to eliminate this whole class of bugs.

Change-Id: Ibd2ac3e8ef7ff05081b43c41a2311144c40d9e23