Bug#37437317 : MySQL Server SEGV within: Table_ref::fetch_number_of_rows
Backport Bug#35647759 fix.

When a function is itemized the flags for the arguments are collected
first and *THEN* the arguments are itemized.
This doesn't account for certain Item itemizations that would replace
the item.
For these the flags aren't aggregated and INSTALL COMPONENT fails to
correctly detect and report the unsupported SET value expression types.

Fixed by itemizing the argument first and then collecting the flags.

Change-Id: I90e20da96aaf764c37b1db8c4ef332e85eb90b74