Bug#35889583: Server fails when we execute set of queries Bug#35996409: Failure in Query_expression::is_set_operation() Bug#36404149: Failing the server
Problem is that we attempt to inspect a query expression object that
has been abandoned due to condition elimination. The problem is with
equality operations that are transformed into semi-join conditions,
where we store the left and right parts of the equality in
sj_inner_exprs and sj_outer_exprs.

The fix is to remove the references in the semi-join expression arrays
when removing the corresponding equality condition. A dive within all
join nests of the query block is necessary, due to the fact that we
may have seen multiple transformations that add join nests so far.

Also had to add some code that was developed for the fix
of bug#35710378. This code was not originally needed for 8.0,
but without it there would be several tests using semi-join that would
change plans.

Change-Id: Ida1bc4c1d586733330b8bd63a3f1f51ac094b96f