Bug#37697512 	InnoDB: `page_track.read_only` big test fails on Windows
Symptom:
The `page_track.read_only` MTR test case has been failing on Windows on
`mysqltest: At line 63: Command "force-rmdir" failed with error 1. my_errno=41.`

Root cause:
The test seems to be removing a directory with files being used by the server, which leads to error deleting these on Windows.
The test seems to need to delete it before the server is started again, and to make things easier it is done before doing restart. This does not work on Windows.

Fix:
We are first stopping the server, and only then delete the files, before starting the server again.
Additionally the `restart_innodb_read_only.inc` was rewritten using existing scripts to reduce code duplication.

Change-Id: If249119b77e3dffc1ffa10faab285d8bd2c057a9