Bug#37397306 - looks like not use pruning when it inserts with now() function in 8.0.
Issue:
All partitions are being retrieved when inserting the now() value
into a non-partition key column in a partition table.

Analysis:
In case of an INSERT query partition pruning takes place in the
prepare stage only. now() was changed from being const to
const_for_execution due to which pruning cannot take place as it
requires all the values in the INSERT to be const.

Fix:
Enable partition pruning to be performed at the execution stage.

Change-Id: I68b7d9d3d880c873a880991d564eca1975900dd0