Bug#36835161: mysql_command_services does not work with x-protocol
The command_services suite uses the sql server sessions as an underlying
mechanism.
This was never supposed to work recursively: i.e. a command using
sessions executed in a session should fail.

Added a check to the command service to ensure it's not usable in such
a context.
Fixed some error handling in the test UDF.

Change-Id: Idb8e4793a3363a1aed6c6b8789445dec8ac462fd