Bug#35625769 Join optimizer SEGV
If the system variable MAX_JOIN_SIZE is set, the optimizer choses
another code path than normal in Query_exression::optimize - it calls
EstimateRowAccesses and compares it with the maximum allowed in
MAX_JOIN_SIZE and gives an error if too large.

Now, in this query, EstimateRowAccesses arrives in
EstimateRowAccessesInItem and the item in question is a
Item_singlerow_subselect. Its query_expression doesn't have a value
set for its root_access_path, so we next call EstimateRowAccesses with
a path argument taken from
Item_singlerow_subselect::root_access_path(), which is always
nullptr. This eventually leads to the issue, when trying to access the
empty path.

Solution: skip calling EstimateRowAccesses is the path is nullptr.

For the hypergraph optimizer, we don't get into this problem because
Query_expression::m_root_access_path is always set, cf. this code:

  if (thd->lex->using_hypergraph_optimizer()) {
    // The hypergraph optimizer also wants all subselect items to be optimized,
    // so that it has cost information to attach to filter nodes.
    for (Query_expression *unit ...)  {
          unit->optimize(thd, ...)  <-- leads to creation of access path for QE
        return true;
    }
    :
  }

so in EstimateRowAccessesInItem:
      if (qe->root_access_path() != nullptr) {
        path = qe->root_access_path();   <-------
      } else {
        path = qe->item->root_access_path();
      }

we get a non-null path from QE's m_root_access_path. This is not
always so for the old optimizer.

Change-Id: I7966d347900014925021fd74f869dee52df8f7ab