Bug#37024069: Issues with strcmp when binlog suffix number exceeds six digits [8.0 patch]
Problem:

In semi-sync replication when the binary log suffix exceeds six digits(.999999)
and becomes binlog-name.1000000 the replication moves from semi-sync to async.

Analysis:

The binlog file naming convention in MySQL utilizes the prefix 'binlog-name.' followed
by a suffix number. When the suffix number is within six digits, it is zero-padded to
maintain the length (e.g.,binlog-name.000001). However, once the suffix exceeds six digits,
the number is not trimmed or padded, resulting in names such as binlog-name.1000000 (1 followed by six zeros).
During replication, the ActiveTranx::compare function, which is implemented using strcmp,
compares the binlog positions. Since strcmp only returns the relationship of the first
non-matching character, it may incorrectly determine that mysql-bin.1000000 comes before mysql-bin.999999.

This incorrect comparison can lead to source no longer sending "confirmation-needed" data
packets to the replica or waiting for confirmations from the replica. As the source stops
waiting for replica's acknowledgments, semi-synchronous replication effectively degrades to
asynchronous replication.

Fix:

In the ActiveTranx::compare function, when comparing the  binlog sequence
numbers, the length of the binlog filenames is compared first; if the lengths
are identical, then comparison is made using strcmp.

Change-Id: Ic6283781706fbe06ad95fe90e6e2988efdbb88c4