Bug#36890891: If SQL in constraint fails, table becomes unopenable / unfixable
When SQL functions are improved, they may throw SQL errors in situations
where they previously did not. If this happens in a table's constraints,
default expressions, partitioning, or virtual columns, the table can not
be opened. This will prevent both analysing the problem (e.g. with SHOW
CREATE TABLE) and addressing it (e.g. with ALTER TABLE ... DROP ...).

* Post-push fix: Fixing interplay with some other (lctn) test cases
                 that also rely on hand-crafted databases.

Change-Id: I69b800a5b0e007894e82cf3065e582d4438d261a