Bug#37687485 ndb_restore : 626 Tuple did not exist
Fix error handling in ndb_restore data restore transaction
on callback.

If one of the MT-restore threads detects a permanent error on
one of the data restore transactions then it should call
set_fatal_error() to make sure that all threads stop, and
the ndb_restore invocation reports an error.

Change-Id: Ifc6a5d759d25a94f4646566f7b4df1bfd71d0319