Bug#37886686 pending_timer cleanup may lead to use-after-free
ASAN reports:

  AddressSanitizer: heap-use-after-free mysql/harness/net_ts/io_context.h:525:9 in
  net::io_context::timer_queue<net::basic_waitable_timer<std::chrono::_V2::steady_clock,
  net::wait_traits<std::chrono::_V2::steady_clock>>>::~timer_queue()

if a timer is still pending at shutdown of the io-context.

Background
==========

At shutdown of a io_context:

1. the pending_timers_ multimap gets destructed which ...
2. destructs each element and unlinks each multimap-node.

If the pending_timer is the last owner Timer, it will try to .cancel()
the timer and remove the Timer from the pending_timers ... which is
currently getting destructed.

Change
======

extract each pending-timer from the "pending-timers" before
destructing.

That way the Timer.cancel() will not find pending-timer in the
pending_timers_ and not try to destroy it. No double double destruct.

Change-Id: Ibfc50d7390ad40c46e75a51e9c2f77805401e118