Bug#28754089 	SPURIOUS CYCLE TIMER NOT AVAILABLE WARNINGS ON SOLARIS
Symptom:
`[Warning] [MY-011037] [Server] The CYCLE timer is not available. WAIT events in the performance_schema will not be timed.`
error messages are visible on SPARC machines during tests.

Root-cause:
On SPARC the tick reaodouts have different values on different sockets.
We are measuring the tick frequency in a 200us window, if we happen to have a context switch to other core, we may have a totally incomparable value reported.
We measure the frequency twice, and get the lowest reading.
If the measured frequency is >= 1THz, i.e. tick is less than picosecond, the resolution in picoseconds is zero, which is used as an error value, and the MY-011037 is reported.
The wrong readout happen sometimes when the CPU is heavily loaded, so it is just a matter of time to get two such cases consecutively, leading to lowest of two readouts to be still incorrectly high.

Fix:
In `my_timer_init_frequency()` we detect if the tick value decreased and if so, we return 0 for incorrect readout.
In `my_timer_init()` we run `my_timer_init_frequency()` up to 10 times, waiting for 3 readouts that are non-zero and less than 1THz, as higher values would be useless for expressing tick resolution in picoseconds.

Change-Id: Iaabd28291a45e6cd4fd4d287fc1db9d5c8328ed4